requirements.txt
streamlit
gspread
google-auth
pandas
numpy
yfinance
youtube-transcript-api
scipy
openpyxl
matplotlib
seaborn
plotly
google-genai
pydantic
finnhub-python
fredapi
requests
python-dotenv
schwab-py>=1.5
httpx>=0.27
google-cloud-storage>=2.18
rich>=13.7.0
typer>=0.12.0
pyyaml>=6.0
gspread-formatting>=1.2.0
ruamel.yaml>=0.18.0
